+++ code/system.ui/dom/src/web.Keyboard/Keyboard.mts
@@ -1,7 +1,7 @@
 import { KeyboardMonitor as Monitor } from './KeyboardMonitor.mjs';
 import { KeyListener } from './KeyListener.mjs';
 import { Match } from './Match.mjs';
-import { Util } from './util.mjs';
+import { Util } from './Util.mjs';
 
 /**
  * Tools for working with a keyboard-input device.
