+++ code/system.ui/dom/src/web.Keyboard/KeyboardMonitor.mts
@@ -1,6 +1,6 @@
 import { Match } from './Match.mjs';
 import { DEFAULTS, R, rx, type t } from './common';
-import { Util } from './util.mjs';
+import { Util } from './Util.mjs';
 
 let _isListening = false;
 let _state: t.KeyboardState = R.clone(DEFAULTS.state);
